Search Endpoints (#681)
* User search

* Track + Playlist search

* PR polish
