+++ discovery-provider/src/api/v1/users.py
@@ -3,9 +3,11 @@ from src.api.v1.models.common import favorite
 from src.api.v1.models.users import user_model
 from src.queries.get_saves import get_saves
 from src.queries.get_users import get_users
+from src.queries.query_helpers import get_current_user_id
+from src.queries.search_queries import SearchKind, search
 from flask_restx import Resource, Namespace, fields, reqparse
 from src.queries.get_tracks import get_tracks
-from src.api.v1.helpers import abort_not_found, decode_with_abort, extend_favorite, extend_track, extend_user, make_response, success_response
+from src.api.v1.helpers import abort_not_found, decode_with_abort, extend_favorite, extend_track, extend_user, make_response, search_parser, success_response
 from .models.tracks import track
 
 logger = logging.getLogger(__name__)
@@ -51,13 +53,24 @@ class FavoritedTracks(Resource):
         favorites = list(map(extend_favorite, favorites))
         return success_response(favorites)
 
-# Still WIP!
-search_parser = reqparse.RequestParser()
-search_parser.add_argument('query', required=True)
+user_search_result = make_response("user_search", ns, fields.List(fields.Nested(user_model)))
+
 @ns.route("/search")
-class SearchResult(Resource):
+class UserSearchResult(Resource):
+    @ns.marshal_with(user_search_result)
     @ns.expect(search_parser)
     def get(self):
+        """Seach for a user"""
         args = search_parser.parse_args()
         query = args["query"]
-        return query, 200
+        search_args = {
+            "query": query,
+            "kind": SearchKind.users.name,
+            "is_auto_complete": False,
+            "current_user_id": None,
+            "with_users": True
+        }
+        response = search(search_args)
+        users = response["users"]
+        users = list(map(extend_user, users))
+        return success_response(users)
